Update patch set 3
Patch Set 3: Code-Review+1

(5 comments)

This is good, it's in line with the evolved architecture.

I'll go through it with you in more detail next week. It seems most of the generated code is in the pdp-client module and the pap-service rest is what outside uses use to talk to the PAP?

Patch-set: 3
Label: Code-Review=+1