Add code to publish to PDP
Added code to take an arbitrary object and send it to PDPs using a
topic sink.  Does not include the code or properties needed to
configure the topic.

Fixed method name - test case does not cover what the method name
implied.
Fixed some comments.

Add test to verify that the PdpClient works with a real TopicSink.

Moved coder classes to policy/common.

Change-Id: I4dec746b07d384c5d9d1449ca91fa39a4f680260
Issue-ID: POLICY-1444
Signed-off-by: Jim Hahn <jrh3@att.com>