Update patch set 2
Patch Set 2:

Changes look good, wondering though how it will be used at runtime.   For example in the context of a healthcheck, I would think the request placed in the bus would contain a request Id, and in theory, all PDPs would respond with the same requestId (but different pdp names).   Thoughts?

Patch-set: 2