Update patch set 2
Patch Set 2:

> > Trying to summarize with regards to this review.   Can you let me
 > > know if my recollection is correct?
 > >
 > > 1. we said in the scrum that the packages *.model.* should be
 > moved
 > > to policy/models repo.
 > >
 > > 2.  I guess we need to come to an agreement in terms of sync and
 > > async as we are discussing by email, so that's still a little in
 > > flux although we are getting near.   If we go all async, wouldn't
 > > "*Response*" objects would go away?
 > >
 > > Feedback?
 > 
 > Jorge:
 > #1 is correct - Ram has already pushed a review for that.
 > #2 If we go async for this particular REST API, then I agree, it
 > seems like the standard HTTP status code & message are sufficient -
 > no need to define a response object.  Or do we have to define
 > something, even if it's just an empty json object? (I'm still
 > learning the nuances of REST.)

As per the specs, we have method to set message & status code in case of error. But in case of success, we can only set the status code as the api for setting message is deprecated. So if we want to send some metadata/info back then we need response object.
Also, it depends upon use case as well.

HTTP Spec says no need to send json response always, use codes instead. Whereas JSON API spec says always send json response.

Hence, if we align our API's to the standard HTTP status codes then we can remove the empty response object. Otherwise we can use it as a document to refer the meaning of the response code/any metadata info. 
Choice is in our hand!

Patch-set: 2