Update patch set 4
Patch Set 4:

> what if we successfully send to some PDPs but not all of them? 
 > Return success?  Return partial success?  I assume we should
 > continue to send to the other PDPs.

I think since we PDP-messages are placed in DMaaP, it will send to all that are listening in the PDP-PAP topic.  It's just one "send" call.

PAP should eventually find out who received it and is operating with it as the PDP-STATUS messages should be sent at regular periodic intervals.

Patch-set: 4
Label: Code-Review=0