Update patch set 6
Patch Set 6:

> Jim, I am approving it but have few points to discuss like json +
 > properties both for configuration, new parameter added to json but
 > no class added, CommonTestData changes etc.
 > May be its better to have a quick zoom call tomorrow and go through
 > these. Is that ok?

Ram: yes, call tomorrow is fine.  1) Regarding json & properties files: given the timeframe, it's too much work to create classes for topic configuration as existing code expects property files.  Doesn't appear to be an easy way to add something like a Properties object to the existing PapParameter class.  2) Regarding new parameter added to the json, I think that was in a previous patch; that was removed in patch #6.

Patch-set: 6