Add PDP communication infrastructure
Includes PDP communication classes, as well as code to push UPDATE
and STATE-CHANGE requests to the PDPs. Does NOT include yet junit code.

Replaced singleton methods with Registry.

Modified constructors to take parameters instead of using Registry,
where possible.  This makes configuration more obvious for junit tests.

Fleshed out PDP deploy provider.  Added JPA code in the provider.  They
will have to be moved or replaced when it becomes available in the JPA
provider layer.  However, the JPA provider layer must either provide a
hook to get the entity manager, or it must accept an entity manager from
the deploy provider, because the deploy provider wants to control the
transaction.

Change-Id: I5b2ae4bb81e729e1b11ec7550b1f88e7b554ce02
Issue-ID: POLICY-1542
Signed-off-by: Jim Hahn <jrh3@att.com>