Update patch set 5
Patch Set 5:

> > (2 comments)
 > 
 > If that's the case, then I think we should modify the
 > PolicyIdentXxx classes to NOT be subclassed from PfConceptKey, but
 > to be simple POJOs, and then modify these classes to use those.

That would also mean that the DAO Provider classes would have to be modified to work with the Policy internal classes instead of the external classes - the interface would have to import from a different place.  Furthermore, if PAP is the only thing that manages the DB for PdpGroups, then it seems to me that those classes should go in policy/pap instead of policy/models, and the corresponding DaoImpl would go there.

(It also seems like the Tosca classes should all come in two flavors then, too - one for external and one for internal, the latter based on ToscaEntityType.)

Patch-set: 5