Added internal classes to mirror models-pap classes
Instead of creating mapper classes, decided to put the conversion methods
within the classes, themselves.  This makes it less likely that someone
will forget to update the conversion method if fields are added or
removed.  Unfortunately, this meant subclassing a few of the models-pdp
classes (e.g., PolicyIdentXxx), to add the required conversion methods.
On the positive side, it means that PAP code need only use the classes
from "internal"; classes from models-pdp can be ignored.
Renamed some of the variables to make it less likely to confuse them.
Modified copyTo() methods to return the target object.

Corrected a few IdentXxx tests.

Change-Id: Ic5aa29fb46f93a90f5b97ea0fd4f6143ed7d5a32
Issue-ID: POLICY-1542
Signed-off-by: Jim Hahn <jrh3@att.com>