Update patch set 11
Patch Set 11:

I am ok with this. It's just that Liam is also planning to bring 2 sets of models - authorative (with basic pojo structure) & persistence (with JPA implementation). Also the mapping between these 2 sets. In that case, we will not need to maintain 1 more set in policy-models-pap and the corresponding Mapper which this review is bringing in.
I am ok if we want to bring it in now and delete later.

Patch-set: 11