Update patch set 11
Patch Set 11:

> I am ok with this. It's just that Liam is also planning to bring 2
 > sets of models - authorative (with basic pojo structure) &
 > persistence (with JPA implementation). Also the mapping between
 > these 2 sets. In that case, we will not need to maintain 1 more set
 > in policy-models-pap and the corresponding Mapper which this review
 > is bringing in.
 > I am ok if we want to bring it in now and delete later.

The "authorative" model for this will be models-pdp (or will replace models-pdp).  However, the external REST API will not use the "authorative" model; it will uses models-pap, thus a mapper will still be needed.  Or would you prefer we eliminate the mapper and add toAuthorative/fromAuthorative methods to the classes in models-pap?

Patch-set: 11