Remove PDP from all active groups
The code to disable a PDP used the group name, extracted from a
prior PdpUpdate request, to identify the group from which the PDP
should be removed.  This is not sufficient, as the requests tracker
may have never seen an update request.  Therefore, the code was
modified to simply remove the PDP from all active groups.  Also
made removeFromGroups(pdpName) public so that it can be used by
other PAP code.

Change-Id: Iedec88cb23e586944563dcc5ac82ff3b9f122f8c
Issue-ID: POLICY-1542
Signed-off-by: Jim Hahn <jrh3@att.com>