Update patch set 2
Patch Set 2: Code-Review+1

This is the "undeploy" code - junits will be provided in a future review.

Note: so-as not to change the swagger API that we gave to CLAMP, we will not be using a POST operation to perform the simple undeploy operation; it will use DELETE, instead.

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1