Update patch set 4
Patch Set 4:

Hi Liam, This is great work and it would be extremely useful for all components, but I am of the thinking that policy/pap is not the right house for it, as the functionality it wouldn't be specific to PAP but it is more a global reusable library/service.   I would even see may be other projects would want to use it for their integration tests.   For example the gson utilities that Jim worked on they were placed in common, also under the assumption that is the best place if other projects want to capitalize on it.

Right now, there are simulators in models (policy/models/tree/models-interactions/model-simulators), which I think, is not the best place neither for them.   It seems to me that policy/common is the best house if it is to be kept within the policy realm.

Right now, none of the PDPs have direct dependencies with any of the PAP artifacts, just with common, and models repos and I have reservations on breaking that pattern.

Patch-set: 4
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0