Sonar fixes for PAP "duplicate code"
Extracted common code from "delete" controller.
Extracted common code from "deploy" controller.
Not sure if this will actually fix the sonar issue or not.  If it's
really complaining about the code, then this should fix it.  On
the other hand, if it's complaining about the annotations, then
I don't think there's a way around it.

Change-Id: I7c68393f3fc1b9719987d850582bd8ed335b9663
Issue-ID: POLICY-1711
Signed-off-by: Jim Hahn <jrh3@att.com>