Adding algorithm for Pdp registration
1) Adding algorithm for improving Pdp registration. The flow is:
When a new Pdp comes for registration, PAP will shortlist a list of
PdpGroups based on the supportedPolicyTypes reported by PdpStatus
message and the supportedPolicyTypes in the PdpSubGroup.
Then PAP will find the PdpGroup which needs the Pdp most in two ways:
1) If there is a subgroup with no Pdp's then that takes the priority.
2) Otherwise the Pdp is allocated to the PdpGroup with highest
difference value of desiredInstanceCount - currentInstanceCount.

2) Changing the preconfigured TosaPolicyType for apex to avoid
interference with legacy operational policies of drools.

Change-Id: I77584c3c030bc27903c11689acaca33148e27e3c
Issue-ID: POLICY-1630
Signed-off-by: ramverma <ram.krishna.verma@est.tech>