Use group from cache even with new policy type
When multiple policies are deployed to the same group in a single
request, and the policies have different policy types, only the
changes associated with the last policy are kept.  Updated the
policy type lookup to use cached groups, when available.
This still needs a junit to verify the fix, but that can come
later.
Also added more logging.

Change-Id: Ieaf866da504b167c884bf53f88aa8cb9e9b5e32a
Issue-ID: POLICY-1761
Signed-off-by: Jim Hahn <jrh3@att.com>