Return 200 when policy successfully undeployed
The PAP Undeploy CSIT was failing, because it was returning 400.
The code was counting PDP UPDATE requests to determine whether or
not the policy had been found.  However, if a subgroup currently
has no PDPs assigned (as is the case with the CSIT), then there
will be no PDP UPDATEs, thus it will think that the policy was
not found and return 400.  Modified the code to count updates to
the DB instead.

Change-Id: I80cdcc738e07c418e0a7184284e7a228e3a6014d
Issue-ID: POLICY-1758
Signed-off-by: Jim Hahn <jrh3@att.com>