Update patch set 1
Patch Set 1:

If we can discuss the GUI code today Liam and Ram that would be great. My concern is when we add Javascript it adds to the codebase and gets counted in sonar code coverage.

I'd like to make sure we have the tools in place that support coverage for Javascript. It can be tricky from what I investigated with the policy/engine project.

Also, the issue of portal integration and uses common look-and-feel.

Patch-set: 1