Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > If we can discuss the GUI code today Liam and Ram that would be great. My concern is when we add Javascript it adds to the codebase and gets counted in sonar code coverage.
> > 
> > I'd like to make sure we have the tools in place that support coverage for Javascript. It can be tricky from what I investigated with the policy/engine project.
> > 
> > Also, the issue of portal integration and uses common look-and-feel.
> 
> SonarJS is released as a plugin of SonarQube and SonarCloud last month. https://www.sonarsource.com/products/codeanalyzers/sonarjs.html  https://docs.sonarqube.org/latest/analysis/languages/javascript/
> Could we integrate it into onap sonarQube? 
> 
> Currently we are creating the client module in PAP to host all the UI code as it suits the most as of now. If decided in future to have a separate repo for UI then we can move the module there. 
> 
> Considerring about portal integration and common style, should I consider it as part of https://jira.onap.org/browse/POLICY-1751 requirements? @Pam, Ram and Liam.

The LF has already configured the ONAP SonarCloud to report Javascript coverage. What we need to do is configure this repository to run tests for Javascript so the coverage gets reported.

In my discussion with Liam and Ram last week, this code is similar to what is in apex so we can proceed forward as that seems to not be reported (yet).

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1