Update patch set 3
Patch Set 3:

The terminology is confusing to me, specifically the word "client", I find it misleading, as in essence it is a plain html web server, monitor gui, or something of that sort I think will make it more clear.   I would like to hear what folks think about it.

Another question is how this functionality will be used, it will run into its own component?

As a note, we owned this repo:  policy/gui .. which would have seem a good candiate to hold all kind of ui work.

Patch-set: 3