Policy notifications appear to be reversed
Modified the code as follows:
Until all PDPs have responded (or been removed), no notifications for
their associated policies will be generated.  However, once a response
has been received for every PDP, then a notification is generated.
For deployed policies, the notifications will continue to be generated
as PDPs move back and forth between the success and failed states, as
long as the policy is not undeployed.

For undeployed policies, on the other hand, no additional notifications
will be generated once all PDPs have successfully undeployed the policy.
However, until that point, like with deployed policies, notifications
will continue to be generated as PDPs move back and forth between the
success and failed states.

Added notification verifications in deploy and undeploy end-to-end tests.

Issue-ID: POLICY-2347
Signed-off-by: Jim Hahn <jrh3@att.com>
Change-Id: I7d18dad1a7500e1efb7074e89ddaca4b1d368965