Update patch set 6
Patch Set 6:

Here's what I'd suggest: create the clients in the controller, rather than in the provider.  This avoids re-creating the clients with every REST call.  Just store them in a list - no need to put them in a map, unless you want to generate warnings for redundant entries.  Pass the list to the provider constructor.

Patch-set: 6