Changes to make PAP container crash with non zero exitCode
Make the application to return exitCode as 1 when the main process
terminates due to issues such as:
1) Any basic issue in the startup config file causing error and
terminates the main process, for e.g. a wrong name in parameterGroup.
2) Wrong info specified for database connectivity, say wrong username.
This doesn't impact a case where database isn't ready yet and PAP
container comes up first, such waiting is already handled by
initContainer in the oom chart. PAP pod waits for policy-mariadb pod to
be ready first, then only the main process is started and comes to this
point.
3) An invalid port number is specified and the rest server is not able
to start.

These are the some of the scenarios which already makes the main process
of PAP to terminate, and in these cases, an exception is thrown now which
returns the exitCode as 1.

Change-Id: Ie48726bcb36d7304eb60807fbaabbee8245e9c70
Issue-ID: POLICY-2754
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>