Update Pdp policy status during registration or State change
The Pdp Policy Status is not updated during the pdp heartbeat
handling/registration flow, or during State change.
When a PDP goes down the pdppolicystatus gets cleared. But when the PDP
comes back online, PAP deploys the right policies into it, but the
deployment status was not getting updated.
Similarly, when the State is changed to PASSIVE, the status should be
undeployed, and when the state is changed back to ACTIVE, then the
status should be deployed for any active policies in any PDP.
These issues are fixed here.

Also updating the csit test to check for 202 instead of 200 for the
deployment api response.

Change-Id: I1e05f22df23daaa8da8be1376413e995be2376f3
Issue-ID: POLICY-2526
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>