Update patch set 1
Patch Set 1:

It would be good to document PAP needs for PDPs as part of this user story.   It may be difficult for folks wanting to add new PDP types to the policy framework to figure out what is expected, and it is getting more complicated over time and don't think is documented.   I think as a guiding principle, for easier growth and adoption of the Policy Framework, PAP should make integration of PDPs as easily as possible, and documentation will help.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>