Send pdp-update if PDP response doesn't match DB
Because multiple PAPs can be updating the DB, it's possible that a
pdp-update sent by a PAP does not reflect the latest deployment data
in the DB. To solve that problem, modified code to compare any response
received from a PDP with what's in the DB, potentially generating a new
pdp-update (and/or pdp-state-change).

Also tried to fix a random junit failure in
TestPolicyComponentsHealthCheckControllerV1.

Issue-ID: POLICY-3426
Change-Id: I241994330d7645c0fffe66abc33de67d71d77250
Signed-off-by: Jim Hahn <jrh3@att.com>