Update patch set 4
Patch Set 4: Code-Review+1

This is good Jim.   Something I failed to spot is if this max-age field is configurable, and where it is initialized or defaulted.  I saw in your junit though but not in the main code.   If you can doublecheck all that is right.

Patch-set: 4
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1