Collect PDP stats via a separate subscription
Added a second subscription to the POLICY-PDP-PAP topic on which PDP
statistics are collected.  This topic uses a shared consumer group so
that only one PAP will receive a given message, thus preventing the same
set of statistics from being written to the DB by multiple PAPs.

Also modified the first subscription so that it no longer updates the
statistics table.

Note: collection of statistics is controlled by the topic configuration;
if the topic, POLICY-STATISTICS, exists, then statistics are collected.
Otherwise, they are not.

Issue-ID: POLICY-3405
Change-Id: I81471f82cff6efac5a59612d3f6af0513022806f
Signed-off-by: Jim Hahn <jrh3@att.com>