Update patch set 2
Patch Set 2:

Found that @DateTimeFormatter is an annotation provided by Spring framework. And we are not there in PAP 😊

Moreover, it is not advisable to use Date as query parameters in rest api's as it is very difficult for clients to manage the format. Best alternative I could find is to use String and declare an "easy" format in which the value is expected. If the value is not in the correct format, send the error response back to the client.

Patch-set: 2