Update patch set 10
Patch Set 10:

Sorry for the big one. Most of the files have small changes only, so I hope it won't be too hard to review.
Note: There are still some improvements to make.
Also verified most of the scenarios, and made sure the CSITs are working. Will raise a separate review for the changes in CSIT.

Patch-set: 10
Work-in-progress: false