Update patch set 1
Patch Set 1:

Nice work Adheli, we need this because in Kafka/Strimzi the topic names may have to be in lower case. The work on the tests was particularly tricky.

Just one comment, maybe we should have a unit test class to check that the topic name overrides work, a test that uses a different application.yaml profile that changes the topic names and just check that the changes are picked up. Also maybe test that the topics have their default values if nothing is specified.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>