Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Nice work Adheli, we need this because in Kafka/Strimzi the topic names may have to be in lower case. The work on the tests was particularly tricky.
> 
> Just one comment, maybe we should have a unit test class to check that the topic name overrides work, a test that uses a different application.yaml profile that changes the topic names and just check that the changes are picked up. Also maybe test that the topics have their default values if nothing is specified.

The end to end block of tests are using a different name for each topic and for the group. It's defined on the application-test-e2e.yaml

I can see if would be possible to test the properties with empty values because they are being re-used inside the application.yaml.

Patch-set: 1