Add OpenAPI Swagger 3 spec and API generation
This commit adds the Policy/API OpenAPI (Swagger) document as a source
document under source control.

This OpenAPI document was generated from the existing SpringFox
annotations in the source code, was uploaded to the online Swagger
editor and was converted into OpenAPI 3.0 format. It was then manually
edited to correct the return types and to match the input and putput
types to our Java classes.

The pom.xml is also amended to generate the API interface source doe in
the target/generated-sources directory.

The generated code does NOT replace the existing API code as yet.

Issue-ID: POLICY-4431
Change-Id: Iee1cc6a122deffcd05728e8f34172f8d423ebcef
Signed-off-by: liamfallon <liam.fallon@est.tech>