Use generated PAP interface for Swagger
This commit removes the Swagger V2 annotations on the
Rest Controllers in policy-pap. The OpeApi annotations (Swagger
v3 annotations) are on a generated Java Interface. The code is changed
so that the controller implements that interface.

There are minimal code changes to the controller and the order of the
parameters is switched in some cases.

Issue-ID: POLICY-4404
Change-Id: I060709806a8e5bc04d9873c8d212a5c9fe13c4d4
Signed-off-by: lapentafd <francesco.lapenta@est.tech>